SDC-3007. Update HistogramJson, MeterJson and TimerJson attributes type to double to preserve the fractions when deserializing
Reviewed-on: https://review.streamsets.net/2932
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>

+++ container-common/src/main/java/com/streamsets/datacollector/restapi/bean/MeterJson.java
@@ -21,15 +21,15 @@ package com.streamsets.datacollector.restapi.bean;
 
 public class MeterJson {
   private long count;
-  private long m1_rate;
-  private long m5_rate;
-  private long m15_rate;
-  private long m30_rate;
-  private long h1_rate;
-  private long h6_rate;
-  private long h12_rate;
-  private long h24_rate;
-  private long mean_rate;
+  private double m1_rate;
+  private double m5_rate;
+  private double m15_rate;
+  private double m30_rate;
+  private double h1_rate;
+  private double h6_rate;
+  private double h12_rate;
+  private double h24_rate;
+  private double mean_rate;
   private String units;
 
   public MeterJson() {
@@ -44,75 +44,75 @@ public class MeterJson {
     this.count = count;
   }
 
-  public long getM1_rate() {
+  public double getM1_rate() {
     return m1_rate;
   }
 
-  public void setM1_rate(long m1_rate) {
+  public void setM1_rate(double m1_rate) {
     this.m1_rate = m1_rate;
   }
 
-  public long getM5_rate() {
+  public double getM5_rate() {
     return m5_rate;
   }
 
-  public void setM5_rate(long m5_rate) {
+  public void setM5_rate(double m5_rate) {
     this.m5_rate = m5_rate;
   }
 
-  public long getM15_rate() {
+  public double getM15_rate() {
     return m15_rate;
   }
 
-  public void setM15_rate(long m15_rate) {
+  public void setM15_rate(double m15_rate) {
     this.m15_rate = m15_rate;
   }
 
-  public long getM30_rate() {
+  public double getM30_rate() {
     return m30_rate;
   }
 
-  public void setM30_rate(long m30_rate) {
+  public void setM30_rate(double m30_rate) {
     this.m30_rate = m30_rate;
   }
 
-  public long getH1_rate() {
+  public double getH1_rate() {
     return h1_rate;
   }
 
-  public void setH1_rate(long h1_rate) {
+  public void setH1_rate(double h1_rate) {
     this.h1_rate = h1_rate;
   }
 
-  public long getH6_rate() {
+  public double getH6_rate() {
     return h6_rate;
   }
 
-  public void setH6_rate(long h6_rate) {
+  public void setH6_rate(double h6_rate) {
     this.h6_rate = h6_rate;
   }
 
-  public long getH12_rate() {
+  public double getH12_rate() {
     return h12_rate;
   }
 
-  public void setH12_rate(long h12_rate) {
+  public void setH12_rate(double h12_rate) {
     this.h12_rate = h12_rate;
   }
 
-  public long getH24_rate() {
+  public double getH24_rate() {
     return h24_rate;
   }
 
-  public void setH24_rate(long h24_rate) {
+  public void setH24_rate(double h24_rate) {
     this.h24_rate = h24_rate;
   }
 
-  public long getMean_rate() {
+  public double getMean_rate() {
     return mean_rate;
   }
 
-  public void setMean_rate(long mean_rate) {
+  public void setMean_rate(double mean_rate) {
     this.mean_rate = mean_rate;
   }
